@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.6 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,38 +29,37 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^6.0.2 
+  provider: ^6.0.5
   flutter_barcode_scanner: ^2.0.0
-  qr_code_scanner: ^0.7.0
+  qr_code_scanner: ^1.0.1
   intl: ^0.17.0
-  assets_audio_player: ^3.0.4+1
+  assets_audio_player: ^3.0.6
   flutter_localizations:
     sdk: flutter
-  shared_preferences: ^2.0.13
-  path_provider: ^2.0.9
-  share_plus: ^4.0.4
-  hive: ^2.1.0
+  shared_preferences: ^2.0.20
+  path_provider: ^2.0.14
+  share_plus: ^6.3.1
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  url_launcher: ^6.1.2
-  fluttertoast: ^8.0.9
-  
-
-
+  url_launcher: ^6.1.10
+  fluttertoast: ^8.2.1
+  http: ^0.13.5
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.5
+  font_awesome_flutter: ^10.4.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.1.10
-  hive_generator: ^1.1.2
+  build_runner: ^2.3.3
+  hive_generator: ^2.0.0
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
