@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class LinkToWhatsApp extends StatelessWidget {
@@ -30,7 +31,7 @@ class LinkToWhatsApp extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: const [
           Icon(
-            Icons.whatsapp,
+            FontAwesomeIcons.whatsapp,
             color: Colors.green,
           ),
           SizedBox(
